iRTG initial implementation
Reworked PiGFX .card driver to build using m68k-amigaos-gcc instead of VBCC
Add some iRTG accelerated functionality to rtg.c
Move rtg_enums.h to match other Amiga driver build setups
pigfx020i.card must be used to enable iRTG usage
